Completion of CacheMatrix functions
CacheMatrix is a set of two functions that allows for the caching of
inverse values of a matrix. This is particularly useful for large
matrices where calculating the inverse may be a costly operation and
must be done so repeatedly. See the included comments for more details
on how to use the functions.